requirements.txt
Flask_Migrate==2.5.3
alembic==1.4.2
Flask==1.1.2
pandas==1.3.2
Werkzeug==1.0.1
Flask_SQLAlchemy==2.4.3
SQLAlchemy==1.3.18
python-dotenv==0.19.1
